Clarified use of lists using Collections and fixed code style inconsistency.
The code "new ArrayList<>(2)" may confuse readers because unless
readers are aware that "2" means the internal buffer is initially
a size of 2, they may think this means there are two items in this
list, which is incorrect. In this commit, we replace this with
"Collections.emptyList()" which is more explicit that it is an empty
list.

Furthermore, this simplified the statements "l = new ArrayList<>(); l.add(item);"
into a single statement "l = Collections.singletonList(item)" that also
forbids mutation of l, which informs readers that the list is not modified later.